Fixed paths on stubMenuItem()
stubMenuItem was using bad paths in real world usage. Gave MenuItem a reference to it’s original view and passed that view to stubMenuItem() so it could borrow it’s `cwd` and `base` settings in the creation of the stubItems view.
Updated tests and fixtures to confirm the changes worked

Signed-off-by: John O'Donnell <john@criticalmash.com>